+++ R/add_scenarios.R
@@ -19,6 +19,10 @@
 #' - `bau`: calls \code{\link{scenario_bau}}
 #' - `accelerate`: calls indicator accelerate function.
 #' - `sdg`: calls Sustainable Development Goals (SDG) acceleration function.
+#' - `covid_rapid_return` calls \code{\link{scenario_covid_rapid_return}}
+#' - `covid_delayed_return` calls \code{\link{scenario_covid_delayed_return}}
+#' - `covid_sustained_disruption` calls \code{\link{scenario_covid_sustained_disruption}}
+#' - `covid_never_return` calls \code{\link{scenario_covid_never_return}}
 #' @param ... additional arguments passed to `add_scenario_indicator`
 #' @inheritParams transform_hpop_data
 #' @inheritParams calculate_hpop_billion
@@ -38,7 +42,11 @@ add_scenario <- function(df,
                            "fixed_target_col",
                            "bau",
                            "accelerate",
-                           "sdg"
+                           "sdg",
+                           "covid_rapid_return",
+                           "covid_delayed_return",
+                           "covid_sustained_disruption",
+                           "covid_never_return"
                          ),
                          ind_ids = billion_ind_codes("all"),
                          ind = "ind",
@@ -102,7 +110,11 @@ add_scenario_indicator <- function(df,
                                      "fixed_target_col",
                                      "bau",
                                      "accelerate",
-                                     "sdg"
+                                     "sdg",
+                                     "covid_rapid_return",
+                                     "covid_delayed_return",
+                                     "covid_sustained_disruption",
+                                     "covid_never_return"
                                    ),
                                    indicator,
                                    ind_ids = billion_ind_codes("all"),